Add files via upload
This is a basic version of a map generator I am working on. It generates new spaces one, two, or three at a time outward from any generated "1" space, with the new spaces opposing the original space which generated the "1" space. I also added in extra code to diminish the possibility of some of the maps only having a couple of "1" spaces. 
The next goal is to make it less likely for new "1" spaces to be generated adjacent to pre-existing one spaces. This will hopefully make the maps look more like thin hallways spiraling in on themselves instead of large rooms lumped together.